refactor: Line 생성 테스트 이동
